requirements.txt
flask
requests
flask-socketio
pymongo
pyyaml
scikit-learn
numpy
pandas
joblib
scapy
psutil
